+++ org.hl7.fhir.validation/src/main/java/org/hl7/fhir/validation/cli/CliContext.java
@@ -1,4 +1,4 @@
-package org.hl7.fhir.validation.cliutils;
+package org.hl7.fhir.validation.cli;
 
 import org.hl7.fhir.validation.Validator;
