+++ org.hl7.fhir.validation/src/main/java/org/hl7/fhir/validation/cli/SnomedVersion.java
@@ -1,4 +1,4 @@
-package org.hl7.fhir.validation.cliutils;
+package org.hl7.fhir.validation.cli;
 
 import java.util.Arrays;
 import java.util.Optional;
